[REF] point_of_sale: reorganize files structure
Reorganization of all `hook` files into a specific folder.

task-id: 4290514
