+++ addons/l10n_in_pos/static/src/company_state_dialog/company_state_dialog.js
@@ -2,7 +2,7 @@
 
 import { Dialog } from "@web/core/dialog/dialog";
 import { Component } from "@odoo/owl";
-import { usePos } from "@point_of_sale/app/store/pos_hook";
+import { usePos } from "@point_of_sale/app/hooks/pos_hook";
 
 export class companyStateDialog extends Component {
     static components = { Dialog };
