+++ addons/point_of_sale/static/src/app/components/order_tabs/order_tabs.js
@@ -1,4 +1,4 @@
-import { usePos } from "@point_of_sale/app/store/pos_hook";
+import { usePos } from "@point_of_sale/app/hooks/pos_hook";
 import { useService } from "@web/core/utils/hooks";
 import { Component, useState } from "@odoo/owl";
 import { ListContainer } from "@point_of_sale/app/generic_components/list_container/list_container";
