+++ addons/point_of_sale/static/src/app/components/popups/cash_move_popup/cash_move_popup.js
@@ -2,11 +2,11 @@ import { _t } from "@web/core/l10n/translation";
 import { useService } from "@web/core/utils/hooks";
 import { parseFloat } from "@web/views/fields/parsers";
 import { Component, useState } from "@odoo/owl";
-import { usePos } from "@point_of_sale/app/store/pos_hook";
+import { usePos } from "@point_of_sale/app/hooks/pos_hook";
 
 import { CashMoveReceipt } from "@point_of_sale/app/components/popups/cash_move_popup/cash_move_receipt/cash_move_receipt";
 import { Dialog } from "@web/core/dialog/dialog";
-import { useAsyncLockedMethod } from "@point_of_sale/app/utils/hooks";
+import { useAsyncLockedMethod } from "@point_of_sale/app/hooks/hooks";
 import { Input } from "@point_of_sale/app/generic_components/inputs/input/input";
 
 export class CashMovePopup extends Component {
