+++ addons/point_of_sale/static/src/app/components/popups/money_details_popup/money_details_popup.js
@@ -1,6 +1,6 @@
 import { Dialog } from "@web/core/dialog/dialog";
 import { Component, useState } from "@odoo/owl";
-import { usePos } from "@point_of_sale/app/store/pos_hook";
+import { usePos } from "@point_of_sale/app/hooks/pos_hook";
 import { floatIsZero } from "@web/core/utils/numbers";
 import { NumericInput } from "@point_of_sale/app/generic_components/inputs/numeric_input/numeric_input";
 import { _t } from "@web/core/l10n/translation";
