+++ addons/point_of_sale/static/src/app/components/popups/select_lot_popup/select_lot_popup.js
@@ -1,5 +1,5 @@
 import { Dialog } from "@web/core/dialog/dialog";
-import { useAutoFocusToLast } from "@point_of_sale/app/utils/hooks";
+import { useAutoFocusToLast } from "@point_of_sale/app/hooks/hooks";
 import { Component, useState, useRef } from "@odoo/owl";
 import { EditListInput } from "@point_of_sale/app/components/popups/select_lot_popup/edit_list_input/edit_list_input";
