+++ addons/point_of_sale/static/src/app/debug/debug_widget.js
@@ -4,7 +4,7 @@ import { Transition } from "@web/core/transition";
 import { useBus, useService } from "@web/core/utils/hooks";
 
 import { useRef, useState, Component } from "@odoo/owl";
-import { usePos } from "@point_of_sale/app/store/pos_hook";
+import { usePos } from "@point_of_sale/app/hooks/pos_hook";
 import { serializeDateTime } from "@web/core/l10n/dates";
 import { ConfirmationDialog } from "@web/core/confirmation_dialog/confirmation_dialog";
 import { makeDraggableHook } from "@web/core/utils/draggable_hook_builder_owl";
