+++ addons/point_of_sale/static/src/app/generic_components/list_container/list_container.js
@@ -1,5 +1,5 @@
 import { Component, useEffect, useRef, xml } from "@odoo/owl";
-import { useIsChildLarger } from "@point_of_sale/app/utils/hooks";
+import { useIsChildLarger } from "@point_of_sale/app/hooks/hooks";
 import { useService } from "@web/core/utils/hooks";
 import { Dialog } from "@web/core/dialog/dialog";
