+++ addons/point_of_sale/static/src/app/hooks/barcode_reader_hook.js
