+++ addons/point_of_sale/static/src/app/hooks/hooks.js
