+++ addons/point_of_sale/static/src/app/hooks/pos_hook.js
