+++ addons/point_of_sale/static/src/app/hooks/time_hook.js
