+++ addons/point_of_sale/static/src/app/navbar/cashier_name/cashier_name.js
@@ -1,5 +1,5 @@
 import { Component, useState } from "@odoo/owl";
-import { usePos } from "@point_of_sale/app/store/pos_hook";
+import { usePos } from "@point_of_sale/app/hooks/pos_hook";
 import { useService } from "@web/core/utils/hooks";
 
 // Previously UsernameWidget
