+++ addons/point_of_sale/static/src/app/navbar/navbar.js
@@ -1,4 +1,4 @@
-import { usePos } from "@point_of_sale/app/store/pos_hook";
+import { usePos } from "@point_of_sale/app/hooks/pos_hook";
 import { useService } from "@web/core/utils/hooks";
 import { isDisplayStandalone, isMobileOS } from "@web/core/browser/feature_detection";
