+++ addons/point_of_sale/static/src/app/navbar/proxy_status/proxy_status.js
@@ -1,5 +1,5 @@
 import { Component, useState } from "@odoo/owl";
-import { usePos } from "@point_of_sale/app/store/pos_hook";
+import { usePos } from "@point_of_sale/app/hooks/pos_hook";
 import { useService } from "@web/core/utils/hooks";
 import { _t } from "@web/core/l10n/translation";
