+++ addons/point_of_sale/static/src/app/navbar/sale_details_button/sale_details_button.js
@@ -2,7 +2,7 @@ import { useService } from "@web/core/utils/hooks";
 import { renderToElement } from "@web/core/utils/render";
 import { AlertDialog } from "@web/core/confirmation_dialog/confirmation_dialog";
 import { Component } from "@odoo/owl";
-import { usePos } from "@point_of_sale/app/store/pos_hook";
+import { usePos } from "@point_of_sale/app/hooks/pos_hook";
 
 export async function handleSaleDetails(pos, hardwareProxy, dialog) {
     const saleDetails = await pos.data.call(
