+++ addons/point_of_sale/static/src/app/screens/product_screen/control_buttons/select_partner_button/select_partner_button.js
@@ -1,5 +1,5 @@
 import { Component, useState } from "@odoo/owl";
-import { usePos } from "@point_of_sale/app/store/pos_hook";
+import { usePos } from "@point_of_sale/app/hooks/pos_hook";
 import { useService } from "@web/core/utils/hooks";
 
 export class SelectPartnerButton extends Component {
