+++ addons/point_of_sale/static/src/app/screens/scale_screen/scale_screen.js
@@ -1,5 +1,5 @@
 import { roundPrecision as round_pr } from "@web/core/utils/numbers";
-import { usePos } from "@point_of_sale/app/store/pos_hook";
+import { usePos } from "@point_of_sale/app/hooks/pos_hook";
 import { Component, onMounted, onWillUnmount, useState } from "@odoo/owl";
 import { Dialog } from "@web/core/dialog/dialog";
 import { useService } from "@web/core/utils/hooks";
