+++ addons/pos_event/static/src/app/popup/event_registration_popup/event_registration_popup.js
@@ -1,7 +1,7 @@
 // Part of Odoo. See LICENSE file for full copyright and licensing details.
 import { Dialog } from "@web/core/dialog/dialog";
 import { Component, useState } from "@odoo/owl";
-import { usePos } from "@point_of_sale/app/store/pos_hook";
+import { usePos } from "@point_of_sale/app/hooks/pos_hook";
 import { ProductCard } from "@point_of_sale/app/generic_components/product_card/product_card";
 import { NumericInput } from "@point_of_sale/app/generic_components/inputs/numeric_input/numeric_input";
 import { useService } from "@web/core/utils/hooks";
