+++ addons/pos_loyalty/static/src/overrides/components/product_screen/product_screen.js
@@ -1,5 +1,5 @@
 import { ProductScreen } from "@point_of_sale/app/screens/product_screen/product_screen";
-import { useBarcodeReader } from "@point_of_sale/app/barcode/barcode_reader_hook";
+import { useBarcodeReader } from "@point_of_sale/app/hooks/barcode_reader_hook";
 import { patch } from "@web/core/utils/patch";
 import { useService } from "@web/core/utils/hooks";
