+++ addons/pos_restaurant/static/src/app/bill_screen/bill_screen.js
@@ -1,7 +1,7 @@
 import { Dialog } from "@web/core/dialog/dialog";
 import { OrderReceipt } from "@point_of_sale/app/screens/receipt_screen/receipt/order_receipt";
 import { Component, useState } from "@odoo/owl";
-import { usePos } from "@point_of_sale/app/store/pos_hook";
+import { usePos } from "@point_of_sale/app/hooks/pos_hook";
 import { useService } from "@web/core/utils/hooks";
 
 export class BillScreen extends Component {
