+++ addons/pos_restaurant/static/src/overrides/components/navbar/table_selector/table_selector.js
@@ -1,4 +1,4 @@
-import { usePos } from "@point_of_sale/app/store/pos_hook";
+import { usePos } from "@point_of_sale/app/hooks/pos_hook";
 import { OrderTabs } from "@point_of_sale/app/components/order_tabs/order_tabs";
 import { NumberPopup } from "@point_of_sale/app/components/popups/number_popup/number_popup";
